refix the signed quelling, use the std::vector.size() type (i.e. size_t) since that's what we're comparing to
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@15010 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/ServerPing.h
@@ -33,7 +33,7 @@ class ServerPing
 {
 public:
   ServerPing();
-  ServerPing(const Address& addr, int port, int _samples = 4, double interval = 1, double tms = 1);
+  ServerPing(const Address& addr, int port, size_t _samples = 4, double interval = 1, double tms = 1);
   ~ServerPing();
   void start();
   int calcLag();
@@ -58,9 +58,9 @@ private:
   
   int fd;
   
-  unsigned int recieved;
+  size_t recieved;
   
-  const unsigned int samples;
+  const size_t samples;
   
   double timeout;
   double interval;
